// src/app/shared/models/repo.model.ts
import { License } from "./license.model";
import { Owner } from "./owner.model";
export interface Repo {
  archive_url?: string;
  archived?: false;
  assignees_url?: string;
  blobs_url?: string;
  branches_url?: string;
  clone_url?: string;
  collaborators_url?: string;
  comments_url?: string;
  commits_url?: string;
  compare_url?: string;
  contents_url?: string;
  contributors_url?: string;
  created_at?: string;
  default_branch?: string;
  deployments_url?: string;
  description?: string;
  disabled?: boolean;
  downloads_url?: string;
  events_url?: string;
  fork?: boolean;
  forks?: number;
  forks_count?: number;
  forks_url?: string;
  full_name?: string;
  git_commits_url?: string;
  git_refs_url?: string;
  git_tags_url?: string;
  git_url?: string;
  has_downloads?: boolean;
  has_issues?: boolean;
  has_pages?: boolean;
  has_projects?: boolean;
  has_wiki?: boolean;
  homepage?: string;
  hooks_url?: string;
  html_url?: boolean;
  id?: number;
  issue_comment_url?: string;
  issue_events_url?: string;
  issues_url?: string;
  keys_url?: string;
  labels_url?: string;
  language?: string;
  languages_url?: string;
  license?: License;
  merges_url?: string;
  milestones_url?: string;
  mirror_url?: string | null;
  name?: string;
  node_id?: string;
  notifications_url?: string;
  open_issues?: number;
  open_issues_count?: number;
  owner?: Owner;
  private?: false;
  pulls_url?: string;
  pushed_at?: string;
  releases_url?: string;
  score?: number;
  size?: number;
  ssh_url?: string;
  stargazers_count?: number;
  stargazers_url?: string;
  statuses_url?: string;
  subscribers_url?: string;
  subscription_url?: string;
  svn_url?: string;
  tags_url?: string;
  teams_url?: string;
  trees_url?: string;
  updated_at?: string;
  url?: string;
  watchers?: number;
  watchers_count?: number;

  avatar_url?: string;
}